5 UnitTests passing in "true R2R"
1) In R2R object writer, I changed previous console output
to regular map file logging; in SectionBuilder I improved it
to include the defined symbols that make it super-easy
to correlate R2R runtime code to the various helpers and such.

2) I have changed the mechanical check for delayed tables
to an explicit property EmitPrecode as most tables no longer
need emitting precodes because the various method and helper
calls use delayed thunks.

One notable exception is string constant acquisition and perhaps
we might be able to emit a special delayed thunk for it too that
would spare us the need to emit precodes for string constants.

3) I have modified the RvaEmbeddedPointer... to implement the
ISymbolDefinitionNode interface so that the map file can tell
us where exactly the different signatures reside in the PE file.

4) I have changed external method call to proper delayed fixup
so that we don't need precodes for that in the runtime method
table.

5) Added WriteBarrier helper and fixed some previous helpers
that incorrectly returned from the switch without updating
the cache.

6) In CorInfoImpl.cs, there's a weird bit where the interface
doesn't seem to honor indirection access - this used to crash
due to trying to execute a data section.

Thanks

Tomas